# compare meshes, unify signatures, and improve performance