--- a/benchmark/.gitignore
+++ b/benchmark/.gitignore
@@ -1,3 +1,3 @@
 *.json
 *.md
-
+*.mesh