--- a/docs/make.jl
+++ b/docs/make.jl
@@ -38,7 +38,9 @@ makedocs(
         "Visualization" => "visualization.md",
         "Conventions" => "conventions.md",
         "Meshes" => [
-            "Unstructured mesh" => joinpath("unstructured_quad_mesh", "unstructured_quad_mesh.md"),
+            "Tree mesh" => joinpath("meshes", "tree_mesh.md"),
+            "Structured mesh" => joinpath("meshes", "structured_mesh.md"),
+            "Unstructured mesh" => joinpath("meshes", "unstructured_quad_mesh.md"),
         ],
         "Time integration" => "time_integration.md",
         "Callbacks" => "callbacks.md",