--- a/src/callbacks_stage/positivity_zhang_shu_dg1d.jl
+++ b/src/callbacks_stage/positivity_zhang_shu_dg1d.jl
@@ -1,7 +1,6 @@
 
-function limiter_zhang_shu!(u::AbstractArray{<:Any,3},
-                            threshold::Real, variable,
-                            mesh, equations, dg::DGSEM, cache)
+function limiter_zhang_shu!(u, threshold::Real, variable,
+                            mesh::AbstractMesh{1}, equations, dg::DGSEM, cache)
   @unpack weights = dg.basis
 
   @threaded for element in eachelement(dg, cache)