--- a/src/callbacks_stage/positivity_zhang_shu_dg2d.jl
+++ b/src/callbacks_stage/positivity_zhang_shu_dg2d.jl
@@ -1,7 +1,6 @@
 
-function limiter_zhang_shu!(u::AbstractArray{<:Any,4},
-                            threshold::Real, variable,
-                            mesh, equations, dg::DGSEM, cache)
+function limiter_zhang_shu!(u, threshold::Real, variable,
+                            mesh::AbstractMesh{2}, equations, dg::DGSEM, cache)
   @unpack weights = dg.basis
 
   @threaded for element in eachelement(dg, cache)