--- a/src/callbacks_step/analysis_dg1d.jl
+++ b/src/callbacks_step/analysis_dg1d.jl
@@ -13,20 +13,20 @@ function create_cache_analysis(analyzer, mesh::TreeMesh,
 end
 
 
-function create_cache_analysis(analyzer, mesh::CurvedMesh, equations::AbstractEquations{1}, 
+function create_cache_analysis(analyzer, mesh::CurvedMesh, equations::AbstractEquations{1},
                                dg::DG, cache, RealT, uEltype)
   # pre-allocate buffers
   u_local = zeros(uEltype,
                   nvariables(equations), nnodes(analyzer))
   x_local = zeros(RealT,
                   ndims(equations), nnodes(analyzer))
-  jacobian_local = zeros(RealT, 
+  jacobian_local = zeros(RealT,
                          nnodes(analyzer))
   return (; u_local, x_local, jacobian_local)
 end
 
 
-function calc_error_norms(func, u::AbstractArray{<:Any,3}, t, analyzer,
+function calc_error_norms(func, u, t, analyzer,
                           mesh::CurvedMesh{1}, equations, initial_condition,
                           dg::DGSEM, cache, cache_analysis)
   @unpack vandermonde, weights = analyzer
@@ -64,7 +64,7 @@ function calc_error_norms(func, u::AbstractArray{<:Any,3}, t, analyzer,
 end
 
 
-function calc_error_norms(func, u::AbstractArray{<:Any,3}, t, analyzer,
+function calc_error_norms(func, u, t, analyzer,
                           mesh::TreeMesh{1}, equations, initial_condition,
                           dg::DGSEM, cache, cache_analysis)
   @unpack vandermonde, weights = analyzer
@@ -100,8 +100,8 @@ function calc_error_norms(func, u::AbstractArray{<:Any,3}, t, analyzer,
 end
 
 
-function integrate_via_indices(func::Func, u::AbstractArray{<:Any,3},
-                               mesh::CurvedMesh, equations, dg::DGSEM, cache,
+function integrate_via_indices(func::Func, u,
+                               mesh::CurvedMesh{1}, equations, dg::DGSEM, cache,
                                args...; normalize=true) where {Func}
   @unpack weights = dg.basis
 
@@ -126,7 +126,7 @@ function integrate_via_indices(func::Func, u::AbstractArray{<:Any,3},
 end
 
 
-function integrate_via_indices(func::Func, u::AbstractArray{<:Any,3},
+function integrate_via_indices(func::Func, u,
                                mesh::TreeMesh{1}, equations, dg::DGSEM, cache,
                                args...; normalize=true) where {Func}
   @unpack weights = dg.basis
@@ -152,8 +152,9 @@ function integrate_via_indices(func::Func, u::AbstractArray{<:Any,3},
 end
 
 
-function integrate(func::Func, u::AbstractArray{<:Any,3},
-                   mesh::Union{TreeMesh{1},CurvedMesh{1}}, equations, dg::DGSEM, cache; normalize=true) where {Func}
+function integrate(func::Func, u,
+                   mesh::Union{TreeMesh{1},CurvedMesh{1}},
+                   equations, dg::DGSEM, cache; normalize=true) where {Func}
   integrate_via_indices(u, mesh, equations, dg, cache; normalize=normalize) do u, i, element, equations, dg
     u_local = get_node_vars(u, equations, dg, i, element)
     return func(u_local, equations)
@@ -161,7 +162,7 @@ function integrate(func::Func, u::AbstractArray{<:Any,3},
 end
 
 
-function analyze(::typeof(entropy_timederivative), du::AbstractArray{<:Any,3}, u, t,
+function analyze(::typeof(entropy_timederivative), du, u, t,
                  mesh::Union{TreeMesh{1},CurvedMesh{1}}, equations, dg::DG, cache)
   # Calculate ∫(∂S/∂u ⋅ ∂u/∂t)dΩ
   integrate_via_indices(u, mesh, equations, dg, cache, du) do u, i, element, equations, dg, du
@@ -171,7 +172,7 @@ function analyze(::typeof(entropy_timederivative), du::AbstractArray{<:Any,3}, u
   end
 end
 
-function analyze(::Val{:l2_divb}, du::AbstractArray{<:Any,3}, u, t,
+function analyze(::Val{:l2_divb}, du, u, t,
                  mesh::TreeMesh{1}, equations::IdealGlmMhdEquations1D,
                  dg::DG, cache)
   integrate_via_indices(u, mesh, equations, dg, cache, dg.basis.derivative_matrix) do u, i, element, equations, dg, derivative_matrix
@@ -184,7 +185,7 @@ function analyze(::Val{:l2_divb}, du::AbstractArray{<:Any,3}, u, t,
   end |> sqrt
 end
 
-function analyze(::Val{:linf_divb}, du::AbstractArray{<:Any,3}, u, t,
+function analyze(::Val{:linf_divb}, du, u, t,
                  mesh::TreeMesh{1}, equations::IdealGlmMhdEquations1D,
                  dg::DG, cache)
   @unpack derivative_matrix, weights = dg.basis