--- a/src/callbacks_step/analysis_dg3d.jl
+++ b/src/callbacks_step/analysis_dg3d.jl
@@ -1,6 +1,6 @@
 
 function create_cache_analysis(analyzer, mesh::TreeMesh{3},
-                               equations::AbstractEquations{3}, dg::DG, cache,
+                               equations, dg::DG, cache,
                                RealT, uEltype)
 
   # pre-allocate buffers
@@ -22,7 +22,7 @@ end
 
 
 function create_cache_analysis(analyzer, mesh::CurvedMesh{3},
-                               equations::AbstractEquations{3}, dg::DG, cache,
+                               equations, dg::DG, cache,
                                RealT, uEltype)
 
   # pre-allocate buffers
@@ -43,13 +43,13 @@ function create_cache_analysis(analyzer, mesh::CurvedMesh{3},
   jacobian_tmp1 = similar(jacobian_local,
                           nnodes(analyzer), nnodes(dg), nnodes(dg))
   jacobian_tmp2 = similar(jacobian_local,
-                          nnodes(analyzer), nnodes(analyzer), nnodes(dg))                                         
+                          nnodes(analyzer), nnodes(analyzer), nnodes(dg))
 
   return (; u_local, u_tmp1, u_tmp2, x_local, x_tmp1, x_tmp2, jacobian_local, jacobian_tmp1, jacobian_tmp2)
 end
 
 
-function calc_error_norms(func, u::AbstractArray{<:Any,5}, t, analyzer,
+function calc_error_norms(func, u, t, analyzer,
                           mesh::TreeMesh{3}, equations, initial_condition,
                           dg::DGSEM, cache, cache_analysis)
   @unpack vandermonde, weights = analyzer
@@ -85,7 +85,7 @@ function calc_error_norms(func, u::AbstractArray{<:Any,5}, t, analyzer,
 end
 
 
-function calc_error_norms(func, u::AbstractArray{<:Any,5}, t, analyzer,
+function calc_error_norms(func, u, t, analyzer,
                           mesh::CurvedMesh{3}, equations, initial_condition,
                           dg::DGSEM, cache, cache_analysis)
   @unpack vandermonde, weights = analyzer
@@ -123,7 +123,7 @@ function calc_error_norms(func, u::AbstractArray{<:Any,5}, t, analyzer,
 end
 
 
-function integrate_via_indices(func::Func, u::AbstractArray{<:Any,5},
+function integrate_via_indices(func::Func, u,
                                mesh::TreeMesh{3}, equations, dg::DGSEM, cache,
                                args...; normalize=true) where {Func}
   @unpack weights = dg.basis
@@ -149,7 +149,7 @@ function integrate_via_indices(func::Func, u::AbstractArray{<:Any,5},
 end
 
 
-function integrate_via_indices(func::Func, u::AbstractArray{<:Any,5},
+function integrate_via_indices(func::Func, u,
                                mesh::CurvedMesh{3}, equations, dg::DGSEM, cache,
                                args...; normalize=true) where {Func}
   @unpack weights = dg.basis
@@ -176,8 +176,9 @@ function integrate_via_indices(func::Func, u::AbstractArray{<:Any,5},
 end
 
 
-function integrate(func::Func, u::AbstractArray{<:Any,5},
-                   mesh::Union{TreeMesh{3},CurvedMesh{3}}, equations, dg::DGSEM, cache; normalize=true) where {Func}
+function integrate(func::Func, u,
+                   mesh::Union{TreeMesh{3},CurvedMesh{3}},
+                   equations, dg::DGSEM, cache; normalize=true) where {Func}
   integrate_via_indices(u, mesh, equations, dg, cache; normalize=normalize) do u, i, j, k, element, equations, dg
     u_local = get_node_vars(u, equations, dg, i, j, k, element)
     return func(u_local, equations)
@@ -185,7 +186,7 @@ function integrate(func::Func, u::AbstractArray{<:Any,5},
 end
 
 
-function analyze(::typeof(entropy_timederivative), du::AbstractArray{<:Any,5}, u, t,
+function analyze(::typeof(entropy_timederivative), du, u, t,
                  mesh::Union{TreeMesh{3},CurvedMesh{3}}, equations, dg::DG, cache)
   # Calculate ∫(∂S/∂u ⋅ ∂u/∂t)dΩ
   integrate_via_indices(u, mesh, equations, dg, cache, du) do u, i, j, k, element, equations, dg, du
@@ -197,7 +198,7 @@ end
 
 
 
-function analyze(::Val{:l2_divb}, du::AbstractArray{<:Any,5}, u, t,
+function analyze(::Val{:l2_divb}, du, u, t,
                  mesh::TreeMesh{3}, equations::IdealGlmMhdEquations3D,
                  dg::DG, cache)
   integrate_via_indices(u, mesh, equations, dg, cache, cache, dg.basis.derivative_matrix) do u, i, j, k, element, equations, dg, cache, derivative_matrix
@@ -212,7 +213,7 @@ function analyze(::Val{:l2_divb}, du::AbstractArray{<:Any,5}, u, t,
   end |> sqrt
 end
 
-function analyze(::Val{:linf_divb}, du::AbstractArray{<:Any,5}, u, t,
+function analyze(::Val{:linf_divb}, du, u, t,
                  mesh::TreeMesh{3}, equations::IdealGlmMhdEquations3D,
                  dg::DG, cache)
   @unpack derivative_matrix, weights = dg.basis