--- a/src/callbacks_step/steady_state_dg1d.jl
+++ b/src/callbacks_step/steady_state_dg1d.jl
@@ -1,6 +1,6 @@
 
-function (steady_state_callback::SteadyStateCallback)(du::AbstractArray{<:Any, 3}, u,
-                                                      mesh, equations, dg::DG, cache)
+function (steady_state_callback::SteadyStateCallback)(du, u, mesh::AbstractMesh{1},
+                                                      equations, dg::DG, cache)
   @unpack abstol, reltol = steady_state_callback
 
   terminate = true