--- a/src/callbacks_step/steady_state_dg3d.jl
+++ b/src/callbacks_step/steady_state_dg3d.jl
@@ -1,6 +1,6 @@
 
-function (steady_state_callback::SteadyStateCallback)(du::AbstractArray{<:Any, 5}, u,
-                                                      mesh, equations, dg::DG, cache)
+function (steady_state_callback::SteadyStateCallback)(du, u, mesh::AbstractMesh{3},
+                                                      equations, dg::DG, cache)
   @unpack abstol, reltol = steady_state_callback
 
   terminate = true