--- a/src/callbacks_step/stepsize_dg1d.jl
+++ b/src/callbacks_step/stepsize_dg1d.jl
@@ -1,5 +1,5 @@
 
-function max_dt(u::AbstractArray{<:Any,3}, t, mesh::TreeMesh{1},
+function max_dt(u, t, mesh::TreeMesh{1},
                 constant_speed::Val{false}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -20,7 +20,7 @@ function max_dt(u::AbstractArray{<:Any,3}, t, mesh::TreeMesh{1},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,3}, t, mesh::TreeMesh{1},
+function max_dt(u, t, mesh::TreeMesh{1},
                 constant_speed::Val{true}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -36,7 +36,7 @@ function max_dt(u::AbstractArray{<:Any,3}, t, mesh::TreeMesh{1},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,3}, t, mesh::CurvedMesh{1},
+function max_dt(u, t, mesh::CurvedMesh{1},
                 constant_speed::Val{false}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -61,7 +61,7 @@ function max_dt(u::AbstractArray{<:Any,3}, t, mesh::CurvedMesh{1},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,3}, t, mesh::CurvedMesh{1},
+function max_dt(u, t, mesh::CurvedMesh{1},
                 constant_speed::Val{true}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection