--- a/src/callbacks_step/stepsize_dg2d.jl
+++ b/src/callbacks_step/stepsize_dg2d.jl
@@ -84,17 +84,17 @@ function max_dt(u::AbstractArray{<:Any,4}, t, mesh::CurvedMesh{2},
       λ1, λ2 = max_abs_speeds(u_node, equations)
 
       # Local speeds transformed to the reference element
-      λ1_transformed = abs(contravariant_vectors[1, 1, i, j, element] * λ1 + 
-                           contravariant_vectors[1, 2, i, j, element] * λ2)
-      λ2_transformed = abs(contravariant_vectors[2, 1, i, j, element] * λ1 + 
-                           contravariant_vectors[2, 2, i, j, element] * λ2)
+      Ja11, Ja12     = get_contravariant_vector(1, contravariant_vectors, i, j, element)
+      λ1_transformed = abs(Ja11 * λ1 + Ja12 * λ2)
+      Ja21, Ja22     = get_contravariant_vector(2, contravariant_vectors, i, j, element)
+      λ2_transformed = abs(Ja21 * λ1 + Ja22 * λ2)
 
       inv_jacobian = inverse_jacobian[i, j, element]
 
       max_λ1 = max(max_λ1, λ1_transformed * inv_jacobian)
       max_λ2 = max(max_λ2, λ2_transformed * inv_jacobian)
     end
-    
+
     max_scaled_speed = max(max_scaled_speed, max_λ1 + max_λ2)
   end
 
@@ -115,10 +115,10 @@ function max_dt(u::AbstractArray{<:Any,4}, t, mesh::CurvedMesh{2},
   for element in eachelement(dg, cache)
     for j in eachnode(dg), i in eachnode(dg)
       # Local speeds transformed to the reference element
-      λ1_transformed = abs(contravariant_vectors[1, 1, i, j, element] * max_λ1 + 
-                           contravariant_vectors[1, 2, i, j, element] * max_λ2)
-      λ2_transformed = abs(contravariant_vectors[2, 1, i, j, element] * max_λ1 + 
-                           contravariant_vectors[2, 2, i, j, element] * max_λ2)
+      Ja11, Ja12     = get_contravariant_vector(1, contravariant_vectors, i, j, element)
+      λ1_transformed = abs(Ja11 * max_λ1 + Ja12 * max_λ2)
+      Ja21, Ja22     = get_contravariant_vector(2, contravariant_vectors, i, j, element)
+      λ2_transformed = abs(Ja21 * max_λ1 + Ja22 * max_λ2)
 
       inv_jacobian = inverse_jacobian[i, j, element]
       max_scaled_speed = max(max_scaled_speed, inv_jacobian * (λ1_transformed + λ2_transformed))