--- a/src/callbacks_step/stepsize_dg3d.jl
+++ b/src/callbacks_step/stepsize_dg3d.jl
@@ -1,5 +1,5 @@
 
-function max_dt(u::AbstractArray{<:Any,5}, t, mesh::TreeMesh{3},
+function max_dt(u, t, mesh::TreeMesh{3},
                 constant_speed::Val{false}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -22,7 +22,7 @@ function max_dt(u::AbstractArray{<:Any,5}, t, mesh::TreeMesh{3},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,5}, t, mesh::TreeMesh{3},
+function max_dt(u, t, mesh::TreeMesh{3},
                 constant_speed::Val{true}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -38,7 +38,7 @@ function max_dt(u::AbstractArray{<:Any,5}, t, mesh::TreeMesh{3},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,5}, t, mesh::CurvedMesh{3},
+function max_dt(u, t, mesh::CurvedMesh{3},
                 constant_speed::Val{false}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection
@@ -73,7 +73,7 @@ function max_dt(u::AbstractArray{<:Any,5}, t, mesh::CurvedMesh{3},
 end
 
 
-function max_dt(u::AbstractArray{<:Any,5}, t, mesh::CurvedMesh{3},
+function max_dt(u, t, mesh::CurvedMesh{3},
                 constant_speed::Val{true}, equations, dg::DG, cache)
   # to avoid a division by zero if the speed vanishes everywhere,
   # e.g. for steady-state linear advection