--- a/src/solvers/solvers.jl
+++ b/src/solvers/solvers.jl
@@ -1,4 +1,4 @@
-include("dg/dg.jl")
+include("dg_tree/dg.jl")
 include("dg_curved/dg.jl")
 include("dg_unstructured_quad/dg.jl")
 include("dg_common.jl")