I used the following code to compare the performance of our different mesh types (running Julia v1.6 with `julia --check-bounds=no --threads=1`).

```julia
julia> using Trixi, BenchmarkTools

julia> begin
           polydeg = 3
           tspan = (0.0, 1.0e-10)

           for elixir in ["benchmark/elixir_2d_euler_vortex_tree.jl",
                          "benchmark/elixir_2d_euler_vortex_structured.jl",
                          "benchmark/elixir_2d_euler_vortex_unstructured.jl",
                          #"benchmark/elixir_3d_euler_source_terms_tree.jl",
                          #"benchmark/elixir_3d_euler_source_terms_structured.jl",
                          ]
               println("#"^80)
               println(elixir)
               redirect_stdout(devnull) do
                   trixi_include(elixir; polydeg, tspan)
               end
               du_ode = similar(ode.u0); du = Trixi.wrap_array(du_ode, semi)
               u_ode = copy(ode.u0); u = Trixi.wrap_array(u_ode, semi)
               display(@benchmark Trixi.rhs!($(du_ode), $(u_ode), $(semi), $(first(tspan))))
               display(@benchmark Trixi.calc_volume_integral!($du, $u, $mesh,
                           $(Trixi.have_nonconservative_terms(equations)),
                           $equations, $(solver.volume_integral), $solver,  $(semi.cache)))
               if mesh isa CurvedMesh
                   display(@benchmark Trixi.calc_interface_flux!($(semi.cache), $u, $mesh, $equations, $solver))
               else
                   display(@benchmark Trixi.prolong2interfaces!($(semi.cache), $u, $mesh, $equations, $solver))
                   display(@benchmark Trixi.calc_interface_flux!($(semi.cache.elements.surface_flux_values),
                               $mesh, $(Trixi.have_nonconservative_terms(equations)),
                               $equations, $solver, $(semi.cache)))
               end
           end
       end
```

Using the changes made in this PR, I get (showing only the benchmarks for `rhs!`)
```
################################################################################
benchmark/elixir_euler_vortex_tree.jl
BenchmarkTools.Trial: 
  memory estimate:  544 bytes
  allocs estimate:  8
  --------------
  minimum time:     97.368 μs (0.00% GC)
  median time:      99.162 μs (0.00% GC)
  mean time:        100.917 μs (0.00% GC)
  maximum time:     250.850 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
################################################################################
benchmark/elixir_euler_vortex_structured.jl
BenchmarkTools.Trial: 
  memory estimate:  544 bytes
  allocs estimate:  8
  --------------
  minimum time:     99.542 μs (0.00% GC)
  median time:      100.742 μs (0.00% GC)
  mean time:        102.777 μs (0.00% GC)
  maximum time:     262.146 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
################################################################################
benchmark/elixir_euler_vortex_unstructured.jl
BenchmarkTools.Trial: 
  memory estimate:  768 bytes
  allocs estimate:  9
  --------------
  minimum time:     130.149 μs (0.00% GC)
  median time:      132.214 μs (0.00% GC)
  mean time:        135.632 μs (0.00% GC)
  maximum time:     307.331 μs (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     1
```

On `main` without the change from `FluxRotated(flux_lax_friedrichs)` to `flux_lax_friedrichs` with a specialization of `flux` to `normal::AbstractVector` instead of `orientation::Integer`, the `CurvedMesh` has an additional overhead of ca. `25 μs`, which is roughly 25% of the total time for `rhs!`.

The `UnstructuredMesh` has the variant of `FluxRotated` hard-coded right now, so it didn't benefit immediately. Hence, we should switch to the same version used for the `CurvedMesh`.

### Results from Rocinante using one thread at ccafced428323f3afb50f1312b5e6c03578ed42c
| ID                                                                      | time ratio                   | memory ratio |
|-------------------------------------------------------------------------|------------------------------|--------------|
| `["2d", "elixir_euler_nonperiodic_curved.jl", "p3_rhs!"]`               | 0.87 (5%) :white_check_mark: |   1.00 (1%)  |
| `["2d", "elixir_euler_nonperiodic_curved.jl", "p7_rhs!"]`               | 0.92 (5%) :white_check_mark: |   1.00 (1%)  |
| `["3d", "elixir_euler_nonperiodic_curved.jl", "p3_rhs!"]`               | 0.66 (5%) :white_check_mark: |   1.00 (1%)  |
| `["3d", "elixir_euler_nonperiodic_curved.jl", "p7_rhs!"]`               | 0.78 (5%) :white_check_mark: |   1.00 (1%)  |

### TODO
- [x] Adapt all 1D, 2D (serial and parallel), and 3D signatures of functions called in `rhs!` to dispatch on the `mesh`
- [x] Implement fluxes dispatching on `normal::AbstractVector` to improve performance
- [x] Get rid of `FluxRotated` in elixirs (but keep one example for CI)
- [x] Add some docs (curvilinear meshes, `FluxRotated`, performance)
- Fix some `TODO: Meshes` (also tracked in #542)

Closes #506, closes #515