@sloede 2021-04-18
> > So what does this PR mean for numerical fluxes in curved meshes: `flux_lax_friedrichs` can directly be used in a curved way, so does `flux`, but all other numerical fluxes would have to be fixed?
> 
> Yes, exactly. Alternatively, other numerical fluxes can use `FluxRotated` for now unless someone needs more performance and implements versions dispatching on `normal::AbstractVector`.

This sounds like a good compromise to me. However, this should definitely be introduced at the next Trixi meeting and preferably also captured in the docs, as it is not 100% clear from reading the code alone